(fix) Fixes circular import issue.
